Setup user feature tests and fix form validations
We now reload the new/edit.js.erb files when there is an error
saving/updating the user. This allows us to get the simple form
validations to appear. I moved some of the bootstrap modal stuff out of
the partial in order to better facilitate the html replacing of the
modal content when a failed save occurs.