Can now edit, create, delete clinics
This follows the same pattern as how the users work. That is, we use
ajax to handle the operations.

A note about testing, unfortunately clinics are persisted across tests
and so we cannot have any tests that add, edit or delete clinics as
these changes will get persisted in the test environment.