Fix and change dashboard
We now show a table on the dashboard page. The **really** tough part about computing counts for the dashboard is dealing with the multiple results we get for visit/test combos. As in we can get a "Pending" result and then a "Negative" result for the same visit/test combo. We want our dashboard to only reflect the **latest** results. This was tricky and required some fancy SQL.